+++ client/src/components/Footer.css
@@ -1,6 +1,6 @@
 .footer {
   position: relative;
-  background-color: white;
+  /* background-color: #1840d9; */
   justify-content: center;
   bottom: 0;
   padding-top: 4rem;
